pyproject.toml
[project]
name = "service-aliyun-dns"
version = "0.1.0"
description = ""
authors = [
    {name = "Colex Jeffers",email = "cfj521@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aliyun-python-sdk-core>=2.14.0",
    "aliyun-python-sdk-alidns>=2.0.7",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
